# Jorje.TheWorld.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Trips API"
  },
  "paths": {
...
    "/api/trips": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsGet",
        "consumes": [],
        "produces": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsPost",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "tripInput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/TripForCreationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsOptions",
        "consumes": [],
        "produces": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/trips/{id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsByIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsByIdPut",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "tripInput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/TripDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsByIdPatch",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocument[TripDTO]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/trips/{id}/stops": {
      "put": {
        "tags": [
          "Trips"
        ],
        "operationId": "ApiTripsByIdStopsPut",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PersonDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "dateOfBirth": {
          "format": "date-time",
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "personImagePath": {
          "type": "string"
        }
      }
    },
...
    "TripForCreationDTO": {
      "type": "object",
      "properties": {}
    },
    "TripDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "stops": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TripStopDTO"
          }
        },
        "persons": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PersonDTO"
          }
        }
      }
    },
    "TripStopDTO": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "arrivalTime": {
          "format": "date-time",
          "type": "string"
        },
        "departureTime": {
          "format": "date-time",
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "trip": {
          "$ref": "#/definitions/TripDTO"
        },
        "stop": {
          "$ref": "#/definitions/StopDTO"
        }
      }
    },
    "StopDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "JsonPatchDocument[TripDTO]": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Operation[TripDTO]"
          },
          "readOnly": true
        }
      }
    },
    "Operation[TripDTO]": {
      "type": "object",
      "properties": {
        "value": {
          "type": "object"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}